requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
firebase-admin==6.4.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.34.0
tavily-python==0.3.3
